Refactor Db ops
